Minor Fehler

+++ tuple.hpp
@@ -62,7 +62,7 @@ public:
     bool monoton();                     ///< Gibt genau dann true zurueck, wenn das Tupel monoton ist.
     bool f(uint32_t i);                 ///< Die Funktion \f$ f_i \f$ fuer \f$ 1 \le i < h \f$.
     bool phi( uint32_t q, uint32_t i);   ///< Die Funtkion \f$ \Phi^q_i \f$.
-    Tuple d_hor( uint32_t i );           ///< Der i-te horizontale Rand \f$ \partial_i^{\prime \prime} 
+    Tuple d_hor( uint8_t i );           ///< Der i-te horizontale Rand \f$ \partial_i^{\prime \prime}
     
     uint32_t p;  ///< The number of symbols \f$ 1 \le p \f$ to be permuted.
 private:
